requirements.txt
langchain>=0.1.14
langchain-openai>=0.0.5
langchain-community>=0.0.34
openai>=1.0.0
tiktoken
python-dotenv
